fix default permissions and umask handling
On linux files are usually created with permissions set to 666+umask,
and folders with 777+umask.

The codebase used to create files with 600 and folders with 700
*unless* a umask override was set in the configuration, then both
files and folders would end up with 777+umask.

This patch addresses this by setting the umask override on process
level; Creating files with permissions set to 666 and folders set to
777 - with either the overridden or the system default umask applied.

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>